Use separate query fragments for ShippingAndPayment Review/Summary
